requirements.txt
av==9.2.0
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
efficientnet-pytorch==0.7.1
fonttools==4.34.4
idna==3.3
imageio==2.21.1
kiwisolver==1.4.4
llvmlite==0.39.0
matplotlib==3.5.2
munch==2.5.0
networkx==2.8.5
numba==0.56.0
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PIMS==0.6.1
pretrainedmodels==0.7.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.0
segmentation-models-pytorch==0.3.0
SimpleITK==2.1.1.2
six==1.16.0
slicerator==1.1.0
tabulate==0.8.10
tifffile==2022.8.3
timm==0.4.12
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
wincertstore==0.2
